♻️ Rename to murmur
Several open-source projects related to secret management are already
called "whisper". In order to make finding this project through search
engines easier, we rename it to "murmur".
