+++ examples/dockerfile/README.md
@@ -1,12 +1,12 @@
 # Example: Dockerfile
 
-This directory contains an example of using whisper inside a Dockerfile.
+This directory contains an example of using murmur inside a Dockerfile.
 
 To run the example:
 
 ```bash
-docker build -t whisper-example .
+docker build -t murmur-example .
 docker run \
     -e SECRET_SAUCE=passthrough:szechuan \
-    whisper-example
+    murmur-example
 ```
