+++ internal/murmur/filter.go
@@ -1,6 +1,6 @@
-package whisper
+package murmur
 
-import "github.com/busser/whisper/internal/whisper/filters/jsonpath"
+import "github.com/busser/murmur/internal/murmur/filters/jsonpath"
 
 // A Filter transforms a value obtained from a secret store into another value
 // based on the given rule.
