+++ internal/murmur/providers/jsonmock/client.go
