+++ internal/murmur/query_test.go
@@ -1,4 +1,4 @@
-package whisper
+package murmur
 
 import (
 	"reflect"
