+++ terraform/README.md
@@ -1,7 +1,7 @@
 # Cloud Infrastucture <!-- omit in toc -->
 
 This directory contains all Terraform code required to provision the cloud
-resources used to test whisper functionality.
+resources used to test murmur functionality.
 
 - [Requirements](#requirements)
 - [Usage](#usage)
